+++ Cargo.toml
@@ -6,9 +6,9 @@ edition = "2021"
 rust-version = "1.65"
 
 [dev-dependencies]
+approx = "0.5.1"
 anyhow = "1.0"
 svg = "0.16.0"
+tracing = "0.1.40"
 
 [dependencies]
-approx = "0.5.1"
-tracing = "0.1.40"
