+++ src/utils.rs
@@ -68,3 +68,64 @@ pub(crate) fn dot(x: (f32, f32), y: (f32, f32)) -> f32 {
 pub fn dist(start: (f32, f32), end: (f32, f32)) -> f32 {
     ((start.0 - end.0).powi(2) + (start.1 - end.1).powi(2)).sqrt()
 }
+
+#[cfg(test)]
+mod test_utils {
+    use crate::utils::{interp, interp2, nearest_point_on_segment, normalize01_32};
+
+    #[test]
+    fn test_normalize_float() {
+        approx::assert_relative_eq!(normalize01_32(1., 2., 1.5), 0.5);
+        approx::assert_relative_eq!(normalize01_32(7., 3., 4.), 0.75); //should also work in reverse
+        approx::assert_relative_eq!(normalize01_32(-1., 1., 2.), 1.);
+        approx::assert_relative_eq!(normalize01_32(1., 1., 1.), 0.);
+        approx::assert_relative_eq!(normalize01_32(1., 1., 0.), 0.);
+        approx::assert_relative_eq!(normalize01_32(1., 1., 2.), 1.);
+    }
+
+    #[test]
+    fn test_inter_float() {
+        approx::assert_relative_eq!(interp(5.0, 10.0, 0.2), 6.0);
+        approx::assert_relative_eq!(interp(10.0, -2.0, 0.75), 1.0);
+        approx::assert_relative_eq!(interp(-1.0, 2.0, -3.0), -1.0);
+        approx::assert_relative_eq!(interp(5.0, 7.0, 20.0), 7.0);
+    }
+
+    #[test]
+    fn test_interp_vec2() {
+        assert_eq!(interp2((1.0, 2.0), (3.0, 5.0), 0.5), (2.0, 3.5));
+        assert_eq!(interp2((-5.0, 5.0), (-15.0, 0.0), 0.4), (-9.0, 3.0));
+        assert_eq!(interp2((7.0, 9.0), (25.0, 30.0), -0.1), (7.0, 9.0));
+        assert_eq!(interp2((12.0, 5.0), (13.0, 14.0), 3.2), (13.0, 14.0));
+    }
+
+    #[test]
+    fn test_nearest_point() {
+        assert_eq!(
+            nearest_point_on_segment((0.0, 0.0), (1.0, 0.0), (0.25, 0.5),),
+            0.25
+        );
+        assert_eq!(
+            nearest_point_on_segment((3.0, 4.0), (5.0, 6.0), (-1.0, -1.0),),
+            0.0
+        );
+        assert_eq!(
+            nearest_point_on_segment((20.0, 10.0), (10.0, 5.0), (2.0, 2.0),),
+            1.0
+        );
+        assert_eq!(
+            nearest_point_on_segment((0.0, 5.0), (5.0, 0.0), (3.0, 3.0),),
+            0.5
+        );
+
+        // degenerate cases
+        assert_eq!(
+            nearest_point_on_segment((0.0, 0.0), (0.0, 0.0), (5.0, 10.0),),
+            0.0
+        );
+        assert_eq!(
+            nearest_point_on_segment((3.0, 7.0), (3.0, 7.0), (0.0, -20.0),),
+            0.0
+        );
+    }
+}
